+++ docker/scripts/change-lambda-image.sh
@@ -79,20 +79,5 @@ echo "updating lambda function ..."
 
 aws lambda wait function-updated --region $aws_region --function-name $AWS_ECR_REPO_NAME
 
-# echo "updating lambda Tags ..."
-# # Update Lamdba configuration to set tag  
-# # Get current environment variables
-# existing_env=$(aws lambda get-function-configuration \
-#   --function-name "$AWS_ECR_REPO_NAME" \
-#   --query 'Environment.Variables' \
-#   --output json)
-
-# # update env with new tag
-# updated_env=$(jq -c --arg TAG "$TAG" '. // {} | .IMAGE_TAG = $TAG' <<< "$existing_env")
-
-# aws lambda update-function-configuration \
-#    --cli-input-json "$(jq -n --arg fn "$AWS_ECR_REPO_NAME" --argjson vars "$updated_env" \
-#      '{FunctionName: $fn, Environment: {Variables: $vars}}')" > /dev/null
-
 echo "success"